Update bioinfo-solutions.md (#2)
* Update bioinfo-solutions.md

Suggested inclusion of Phylopipe - this is the tree generating pipeline used by COG-UK for their analysis and produces the files on which Microreact , CoV-GLUE etc are based. Defaults are set up for COG-UK applications, but lots of flexibility for tree building to be performed elsewhere for other applications e.g. I use it also on just GISAID data. There is also a second pipeline Datapipe (https://github.com/COG-UK/grapevine_nextflow) which does the preprocessing (cleaning, alignment, variant and lineage calling, deduplication) and creates the incoming files for Phylopipe within the COG-UK setup.

* Add datapipe

Added datapipe to section 3 and mention in phylopipe section